fix failed test

+++ packages/tfchain_client/test/balances_test.dart
@@ -62,7 +62,7 @@ void main() {
         AccountInfo? before =
             await setupManager.client.balances.get(address: recipientAddress);
         await setupManager.client.balances.transfer(
-            address: recipientAddress, amount: randomNumber);
+            address: recipientAddress, amount: randomNumber.toDouble());
 
         AccountInfo? after =
             await setupManager.client.balances.get(address: recipientAddress);
